test (WinningLottoValidatorTest): 당첨 숫자 검증 테스트 이동
